refactor(shared)(game): move response DTOs into shared transport package
