+++ apps/shared/src/main/kotlin/shared/models/game/response/BoardSetResponse.kt
@@ -1,4 +1,4 @@
-package at.se2group.backend.dto
+package shared.models.game.response
 
 data class BoardSetResponse(
     val boardSetId: String,
